Created A Customizable Modular Wave System
This wave system I created allows me to choose which enemies I want to spawn in which order. It also has a lot of feature such as being able to decide which enemies give powerups, playing a specific timeline pattern, and choosing to spawn when there are no enemies on screen.